Create basic logic

+++ scr/telegram_bot.py
@@ -11,22 +11,20 @@ logging.basicConfig(
 
 class TelegramBot:
     def __init__(self):
-        self.LIST_MESSAGES = []
+        self.DICT_MESSAGES = {}
         self.bot = telegram.Bot(os.environ["TELEGRAM_TOKEN"])
 
-
-    async def send_message(self) -> None:
-        message = await self.bot.send_message(
-            chat_id=os.environ["CHANNEL"], text="I'm a bot, please talk to me!"
-        )
-        self.LIST_MESSAGES.append(message.message_id)
-
+    async def send_message(self, title, url) -> None:
+        if not self.DICT_MESSAGES.get(title, None):
+            message = await self.bot.send_message(
+                chat_id=os.environ["CHANNEL"], text=f"I'm a bot, please talk to me! {url}"
+            )
+            self.DICT_MESSAGES[title] = message.message_id
 
     async def delete_message(self) -> None:
-        while self.LIST_MESSAGES:
-            message_id = self.LIST_MESSAGES.pop()
-            await self.bot.delete_message(
-                chat_id=os.environ["CHANNEL"], message_id=message_id
-        )
-
-
+        while self.DICT_MESSAGES:
+            message = self.DICT_MESSAGES.popitem()
+            if message:
+                await self.bot.delete_message(
+                    chat_id=os.environ["CHANNEL"], message_id=message[-1]
+            )
